--- shared/src/commonTest/kotlin/vn/luongvo/kmm/survey/test/MockUtil.kt
@@ -1,12 +1,0 @@
-package vn.luongvo.kmm.survey.test
-
-import vn.luongvo.kmm.survey.data.remote.model.response.TokenResponse
-import vn.luongvo.kmm.survey.data.remote.model.response.toToken
-
-val tokenResponse = TokenResponse(
-    tokenType = "tokenType",
-    accessToken = "accessToken",
-    refreshToken = "refreshToken"
-)
-
-val token = tokenResponse.toToken()
